+++ skbio/__init__.py
@@ -13,19 +13,16 @@ from skbio.util import TestRunner
 # Add skbio.io to sys.modules to prevent cycles in our imports
 import skbio.io  # noqa
 # imports included for convenience
-from skbio.sequence import (
-    BiologicalSequence, NucleotideSequence, DNA, DNASequence, RNA, RNASequence,
-    Protein, ProteinSequence)
+from skbio.sequence import Sequence, DNA, RNA, Protein
 from skbio.stats.distance import DistanceMatrix
 from skbio.alignment import (
     local_pairwise_align_ssw, SequenceCollection, Alignment)
-from skbio.tree import (
-    TreeNode, nj)
+from skbio.tree import TreeNode, nj
 from skbio.io import read, write
 
-__all__ = ['BiologicalSequence', 'NucleotideSequence', 'DNA', 'DNASequence',
-           'RNA', 'RNASequence', 'Protein', 'ProteinSequence',
-           'DistanceMatrix', 'local_pairwise_align_ssw', 'SequenceCollection',
+
+__all__ = ['Sequence', 'DNA', 'RNA', 'Protein', 'DistanceMatrix',
+           'local_pairwise_align_ssw', 'SequenceCollection',
            'Alignment', 'TreeNode', 'nj', 'read', 'write']
 
 __credits__ = "https://github.com/biocore/scikit-bio/graphs/contributors"
@@ -35,7 +32,9 @@ mottos = [
     # 03/15/2014
     "It's gonna get weird, bro.",
     # 05/14/2014
-    "no cog yay"
+    "no cog yay",
+    # 03/18/2015
+    "bincount!",
 ]
 motto = mottos[-1]