+++ skbio/util/_exception.py
@@ -12,3 +12,8 @@ from __future__ import absolute_import, division, print_function
 class TestingUtilError(Exception):
     """Raised when an exception is needed to test exception handling."""
     pass
+
+
+class OverrideError(AssertionError):
+    """Raised when a property does not exist in the parent class."""
+    pass